test: avoid nested Cargo in gix-filter driver tests
Concurrent nextest processes could each invoke `cargo run --example arrow` and
race while Cargo replaced the shared example artifact. This made driver tests
intermittently fail when Cargo could not execute the path it had just built.

Declare the Arrow helper as a dedicated test binary, let Cargo build it before
integration tests, and use its `CARGO_BIN_EXE_*` path with repository-standard
shell quoting. Move the helper source into the packaged test-helper layout
without duplicating targets, and remove serial_test now that tests do not
coordinate shared state.
