+++ gix-filter/Cargo.toml
@@ -9,16 +9,33 @@ description = "A crate of the gitoxide project implementing git filters"
 authors = ["Sebastian Thiel <sebastian.thiel@icloud.com>"]
 edition = "2024"
 rust-version = "1.85"
-include = ["/src/**/*", "/LICENSE-*"]
+include = [
+    "/src/**/*",
+    "/LICENSE-*",
+    # Cargo packages declared examples and binaries, and the example includes the test helper.
+    "/examples/arrow.rs",
+    "/tests/helpers/arrow.rs",
+]
 
 [lib]
 doctest = false
+
+[[bin]]
+name = "gix-filter-test-arrow"
+path = "tests/helpers/arrow.rs"
+# The self dev-dependency below enables this only while building tests.
+required-features = ["test-helper"]
+test = false
+bench = false
+doc = false
 
 [features]
 ## Enable support for the SHA-1 hash by forwarding the feature to dependencies.
 sha1 = ["gix-hash/sha1"]
 ## Enable support for the SHA-256 hash by forwarding the feature to dependencies.
 sha256 = ["gix-hash/sha256"]
+## Internal feature for building binaries referenced by integration tests through CARGO_BIN_EXE_*.
+test-helper = []
 
 [dependencies]
 gix-hash = { version = "^0.25.1", path = "../gix-hash" }
@@ -38,12 +55,11 @@ smallvec = "1.15.1"
 
 
 [dev-dependencies]
+gix-filter = { path = ".", features = ["test-helper"] }
 gix-hash = { path = "../gix-hash", features = ["sha1", "sha256"] }
 gix-quote = { path = "../gix-quote" }
 gix-testtools = { path = "../tests/tools" }
 gix-worktree = { path = "../gix-worktree", default-features = false, features = ["attributes"] }
-
-serial_test = { version = "3.4.0", default-features = false }
 
 [package.metadata.docs.rs]
 features = ["sha1"]
