+++ gix-filter/tests/filter/driver.rs
@@ -1,17 +1,11 @@
-static DRIVER: &str = concat!(env!("CARGO"), " run --example arrow");
+use bstr::{BStr, BString};
 
 mod baseline {
-    use serial_test::serial;
+    use crate::driver::driver_path;
 
-    use crate::driver::DRIVER;
-
-    #[serial]
     #[test]
     fn our_implementation_used_by_git() -> crate::Result {
-        let mut exe = DRIVER.to_owned();
-        if cfg!(windows) {
-            exe = exe.replace('\\', "/");
-        }
+        let exe = driver_path().to_string();
         gix_testtools::scripted_fixture_read_only_with_args_single_archive("baseline.sh", [exe])?;
         Ok(())
     }
@@ -61,14 +55,12 @@ mod shutdown {
 pub(crate) mod apply {
     use std::{io::Read, sync::LazyLock};
 
-    use bstr::{BStr, BString, ByteSlice};
+    use crate::driver::{driver_path, shutdown::extract_client};
+    use bstr::{BStr, BString, ByteSlice, ByteVec};
     use gix_filter::{
         Driver, driver,
         driver::{Operation, apply, apply::Delay},
     };
-    use serial_test::serial;
-
-    use crate::driver::{DRIVER, shutdown::extract_client};
 
     fn driver_no_process() -> Driver {
         let mut driver = driver_with_process();
@@ -77,20 +69,20 @@ pub(crate) mod apply {
     }
 
     pub(crate) fn driver_with_process() -> Driver {
-        let mut exe = DRIVER.to_owned();
-        if cfg!(windows) {
-            exe = exe.replace('\\', "/");
-        }
+        let command = |suffix| {
+            let mut command = driver_path();
+            command.push_str(suffix);
+            command
+        };
         Driver {
             name: "arrow".into(),
-            clean: Some((exe.clone() + " clean %f").into()),
-            smudge: Some((exe.clone() + " smudge %f").into()),
-            process: Some((exe + " process").into()),
+            clean: Some(command(" clean %f")),
+            smudge: Some(command(" smudge %f")),
+            process: Some(command(" process")),
             required: true,
         }
     }
 
-    #[serial]
     #[test]
     fn missing_driver_means_no_filter_is_applied() -> crate::Result {
         let mut state = gix_filter::driver::State::default();
@@ -121,7 +113,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn a_crashing_process_can_restart_it() -> crate::Result {
         let mut state = gix_filter::driver::State::default();
@@ -155,7 +146,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn process_status_abort_disables_capability() -> crate::Result {
         let mut state = gix_filter::driver::State::default();
@@ -184,7 +174,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn process_status_strange_shuts_down_process() -> crate::Result {
         let mut state = gix_filter::driver::State::default();
@@ -212,7 +201,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn smudge_and_clean_failure_is_translated_to_observable_error_for_required_drivers() -> crate::Result {
         let mut state = gix_filter::driver::State::default();
@@ -222,7 +210,10 @@ pub(crate) mod apply {
         let mut filtered = state
             .apply(
                 &driver,
-                &mut &b"hello\nthere\n"[..],
+                // The helper intentionally exits before reading stdin. Keep the input empty so
+                // the observable error is the driver failure, not a scheduling-dependent broken
+                // pipe from the background stdin writer.
+                &mut std::io::empty(),
                 driver::Operation::Smudge,
                 context_from_path("do/fail"),
             )?
@@ -234,7 +225,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn smudge_and_clean_failure_means_nothing_if_required_is_false() -> crate::Result {
         let mut state = gix_filter::driver::State::default();
@@ -258,7 +248,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn smudge_and_clean_series() -> crate::Result {
         let mut state = gix_filter::driver::State::default();
@@ -378,7 +367,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn large_file_with_cat_filter_does_not_hang() -> crate::Result {
         // This test reproduces issue #2080 where using `cat` as a filter with a large file
@@ -431,7 +419,6 @@ pub(crate) mod apply {
         Ok(())
     }
 
-    #[serial]
     #[test]
     fn large_file_with_cat_filter_early_drop() -> crate::Result {
         // Test that dropping the reader early doesn't cause issues (thread cleanup)
@@ -505,4 +492,15 @@ pub(crate) mod apply {
             b"cat".into()
         }
     }
+}
+
+fn driver_path() -> BString {
+    fn quote_driver_path(path: &str) -> BString {
+        let path = gix_path::to_unix_separators_on_windows(BStr::new(path));
+        gix_quote::single(path.as_ref())
+    }
+
+    // Cargo builds binary targets before integration tests and provides their absolute paths.
+    // Invoking Cargo recursively here would race with other nextest processes over the same artifact.
+    quote_driver_path(env!("CARGO_BIN_EXE_gix-filter-test-arrow"))
 }
